fix: 故事模式 v2.1.3 编译修复 + spec 收尾
- CommandExecutor: PRIMED_TNT -> TNT（1.21 API）
- RulebookListener: PlayerEditBookEvent.getBook() 不存在，改从玩家手中读取书
- StoryManager: tick 间隔 int -> long 避免 long-to-int 转换
- tasks.md / checklist.md: 勾选已完成项

+++ src/main/java/com/aip/story/StoryManager.java
@@ -47,7 +47,7 @@ public class StoryManager {
         if (initialized) return;
         initialized = true;
 
-        int aerialInterval = plugin.getConfigManager().getAerialTickMs() / 50;  // ms → tick
+        long aerialInterval = plugin.getConfigManager().getAerialTickMs() / 50;  // ms → tick
         if (aerialInterval < 1) aerialInterval = 80;  // 默认 4 秒
 
         aerialTask = Bukkit.getScheduler().runTaskTimer(plugin, this::tickAerialAssault, aerialInterval, aerialInterval);
